Removed numba dependendancy and replaced with pure numpy version, runs slightly faster but takes up more memory.
