+++ requirements.txt
@@ -1,3 +1,2 @@
 numpy>=1.20.1
 scipy>=1.7.1
-numba>=0.53.1
