+++ src/__init__.py
@@ -1,7 +1,7 @@
-from .numba_functions import _array_LUT
 from .utils import *
 from .polytomous_utils import *
 from .synthetic import *
+from .look_up_table import *
 from .latent_ability_distribution import *
 from .ability_methods import *
 from .mml_methods import *
