+++ src/multidimensional/multidimensional_mml_methods.py
@@ -3,10 +3,10 @@ from itertools import repeat, product
 import numpy as np
 from scipy.optimize import fmin_slsqp, fminbound
 
-from girth import (condition_polytomous_response,
+from girth import (condition_polytomous_response,create_beta_LUT,
                    validate_estimation_options)
 
-from girth.utils import create_beta_LUT, INVALID_RESPONSE
+from girth.utils import INVALID_RESPONSE
 from girth.latent_ability_distribution import LatentPDF
 from girth.multidimensional import multidimensional_ability_eap, initial_guess_md
 from girth.polytomous_utils import (_graded_partial_integral_md, _solve_for_constants,
